move test cases into dedicated file
