+++ tests/python/unittest/test_tir_ir_builder.py
@@ -554,147 +554,6 @@ def test_dyn_shared():
         check_target(target)
 
 
-@tvm.testing.requires_gpu
-def test_matmul_dyn_shared():
-    n = 1024
-    A = te.placeholder((n, n), name="A", dtype="float16")
-    B = te.placeholder((n, n), name="B", dtype="float16")
-
-    def syncthread():
-        return tvm.tir.Call(None, "tir.tvm_storage_sync", tvm.runtime.convert(["shared"]))
-
-    def test_matmul_ir(A, B, C):
-        ib = tvm.tir.ir_builder.create()
-        block = 16
-
-        tx = te.thread_axis("threadIdx.x")
-        ty = te.thread_axis("threadIdx.y")
-        bx = te.thread_axis("blockIdx.x")
-        by = te.thread_axis("blockIdx.y")
-        ib.scope_attr(tx, "thread_extent", block)
-        ib.scope_attr(ty, "thread_extent", block)
-        ib.scope_attr(bx, "thread_extent", n / block)
-        ib.scope_attr(by, "thread_extent", n / block)
-
-        A_sh = ib.allocate(A.dtype, (block, block), scope="shared.dyn", name="A_sh")  # fp16
-        B_sh = ib.allocate(B.dtype, (block, block), scope="shared.dyn", name="B_sh")  # fp16
-        # Create a dynamic shared memory for the accumulation.
-        # This is for testing merging dynamic shared memory alloctions with different data type.
-        # In practice, there is no need to allocate a shared memory for C.
-        C_sh = ib.allocate(C.dtype, (block, block), scope="shared.dyn", name="C_sh")  # fp32
-
-        A_ptr = ib.buffer_ptr(A)
-        B_ptr = ib.buffer_ptr(B)
-        C_ptr = ib.buffer_ptr(C)
-
-        C_sh[ty, tx] = 0.0
-
-        with ib.for_range(0, n // block, name="i") as i:
-            A_sh[ty, tx] = A_ptr[by * block + ty, i * block + tx]
-            B_sh[ty, tx] = B_ptr[i * block + ty, bx * block + tx]
-            ib.emit(syncthread())
-
-            with ib.for_range(0, block, name="k") as k:
-                C_sh[ty, tx] += cast(A_sh[ty, k] * B_sh[k, tx], "float32")
-
-            ib.emit(syncthread())
-
-        C_ptr[by * block + ty, bx * block + tx] = C_sh[ty, tx]
-
-        return ib.get()
-
-    C = te.extern(
-        A.shape,
-        [A, B],
-        lambda ins, outs: test_matmul_ir(ins[0], ins[1], outs[0]),
-        name="reduce",
-        dtype="float32",
-    )
-    s = te.create_schedule(C.op)
-
-    def check_target(target):
-        if not tvm.testing.device_enabled(target):
-            return
-
-        fmatmul = tvm.build(s, [A, B, C], target)
-        dev = tvm.device(target, 0)
-
-        size = (n, n)
-        a_np = np.random.uniform(size=size).astype(A.dtype)
-        b_np = np.random.uniform(size=size).astype(B.dtype)
-        a = tvm.nd.array(a_np, dev)
-        b = tvm.nd.array(b_np, dev)
-        c = tvm.nd.array(np.zeros(size, dtype=C.dtype), dev)
-        fmatmul(a, b, c)
-        np_ref = np.dot(a_np.astype("float32"), b_np.astype("float32"))
-        tvm.testing.assert_allclose(c.numpy(), np_ref, 1e-4, 1e-4)
-
-    for target in ["cuda", "nvptx"]:
-        check_target(target)
-
-
-@tvm.testing.requires_gpu
-def test_dyn_shared_vectorized_store():
-    """Test vectorized store into dynamic shared memory"""
-    n = te.size_var("n")
-    A = te.placeholder((n,), name="A", dtype="float16")
-    B = te.placeholder((n,), name="B", dtype="float32")
-
-    def test_device_ir(A, B, C):
-        n = A.shape[0]
-        ib = tvm.tir.ir_builder.create()
-
-        values_per_thread = 4
-        tx = te.thread_axis("threadIdx.x")
-        ib.scope_attr(tx, "thread_extent", tvm.tir.indexdiv(n, values_per_thread))
-
-        A_sh = ib.allocate(A.dtype, (n,), scope="shared.dyn")  # fp16
-        B_sh = ib.allocate(B.dtype, (n,), scope="shared.dyn")  # fp32
-
-        Aptr = ib.buffer_ptr(A)
-        Bptr = ib.buffer_ptr(B)
-        Cptr = ib.buffer_ptr(C)
-
-        with ib.for_range(0, values_per_thread, kind="vectorize") as i:
-            A_sh[tx * values_per_thread + i] = Aptr[tx * values_per_thread + i]
-            B_sh[tx * values_per_thread + i] = Bptr[tx * values_per_thread + i]
-
-        with ib.for_range(0, values_per_thread) as i:
-            Cptr[tx * values_per_thread + i] = (
-                cast(A_sh[tx * values_per_thread + i], "float32") + B_sh[tx * values_per_thread + i]
-            )
-
-        return ib.get()
-
-    C = te.extern(
-        (n,),
-        [A, B],
-        lambda ins, outs: test_device_ir(ins[0], ins[1], outs[0]),
-        name="vadd",
-        dtype="float32",
-    )
-    s = te.create_schedule(C.op)
-
-    def check_target(target):
-        if not tvm.testing.device_enabled(target):
-            return
-
-        fadd = tvm.build(s, [A, B, C], target)
-        dev = tvm.device(target, 0)
-
-        for n in [512, 1024]:
-            a = tvm.nd.array(np.random.uniform(size=n).astype(A.dtype), dev)
-            b = tvm.nd.array(np.random.uniform(size=n).astype(B.dtype), dev)
-            c = tvm.nd.array(np.zeros((n,), dtype=C.dtype), dev)
-            fadd(a, b, c)
-            tvm.testing.assert_allclose(
-                c.numpy(), a.numpy().astype("float32") + b.numpy(), 1e-4, 1e-4
-            )
-
-    for target in ["cuda", "nvptx"]:
-        check_target(target)
-
-
 if __name__ == "__main__":
     test_prefetch()
     test_if()
@@ -706,5 +565,3 @@ if __name__ == "__main__":
     test_while_mandel()
     test_while_binary_search()
     test_dyn_shared()
-    test_matmul_dyn_shared()
-    test_dyn_shared_vectorized_store()
